REFACTOR: Remove unused files and group files by task
